+++ src/main/java/xyz/kohara/stellarity/Stellarity.java
@@ -23,31 +23,35 @@ public class Stellarity implements ModInitializer {
     public static ResourceLocation id(String path) {
         //? if = 1.20.1 {
         return new ResourceLocation(MOD_ID, path);
-        //?} else {
+         //?} else {
         /*return ResourceLocation.fromNamespaceAndPath(MOD_ID, path);
-         *///?}
+        *///?}
 
     }
 
     public static ResourceLocation id(String namespace, String path) {
         //? if = 1.20.1 {
         return new ResourceLocation(namespace, path);
-        //?} else {
+         //?} else {
         /*return ResourceLocation.fromNamespaceAndPath(namespace, path);
-         *///?}
+        *///?}
 
     }
 
     public static ResourceLocation mcId(String path) {
         //? if = 1.20.1 {
         return new ResourceLocation(path);
-        //?} else {
+         //?} else {
         /*return ResourceLocation.withDefaultNamespace(path);
-         *///?}
+        *///?}
     }
 
     public static <T extends Registry<U>, U> ResourceKey<U> key(ResourceKey<T> registry, String path) {
         return ResourceKey.create(registry, id(path));
+    }
+
+    public static <T extends Registry<U>, U> ResourceKey<U> mcKey(ResourceKey<T> registry, String path) {
+        return ResourceKey.create(registry, mcId(path));
     }
 
     @Override
@@ -63,6 +67,7 @@ public class Stellarity implements ModInitializer {
         StellarityCreativeModeTabs.init();
         StellarityEntities.init();
         StellarityBlockEntityTypes.init();
+        StellarityLootTables.init();
         StellarityCriteriaTriggers.init();
         StellarityRecipeTypes.init();
         StellarityRecipeSerializers.init();