+++ src/main/java/xyz/kohara/stellarity/datagen/AdvancementProvider.java
@@ -10,25 +10,28 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.advancements.critereon.*;
 //? } else {
-/*import net.minecraft.advancements.criterion.*;
+/*import net.minecraft.advancements.critereon.*;
 *///? }
 import xyz.kohara.stellarity.Stellarity;
 import xyz.kohara.stellarity.registry.StellarityItems;
 
-import java.util.HashMap;
 import java.util.function.Consumer;
+
+import xyz.kohara.stellarity.registry.advancement_criterion.VoidFishedTrigger;
 //? >= 1.21.1 {
 /*import net.minecraft.core.HolderLookup;
-import xyz.kohara.stellarity.registry.advancement_criterion.VoidFishedTrigger;
 import net.minecraft.world.item.Item;
+
 import java.util.concurrent.CompletableFuture;
 import java.util.List;
 import java.util.Optional;
+
 import net.minecraft.core.registries.Registries;
 *///?} else {
 import net.minecraft.advancements.FrameType;
 import net.minecraft.advancements.CriterionTriggerInstance;
-import xyz.kohara.stellarity.registry.advancement_criterion.VoidFishedTrigger;
+
+import java.util.HashMap;
 //?}
 
 public class AdvancementProvider extends FabricAdvancementProvider {